bevService

+++ services/bevService.js
@@ -3,12 +3,11 @@ export class BEVService {
     constructor(contract) {
         this.contract = contract;
     }
-
-    // TODO: probar la respuesta en formato JSON
+    
     // Respuesta en formato JSON
     response(_status, _data){
-        //var d = JSON.stringify(_data);
-        var r = '{"status":'+_status+',"data":'+_data+'}'
+        var d = JSON.stringify(_data);
+        var r = '{"status":'+_status+',"data":'+ d +'}'
         var o = JSON.parse(r);         
         return o;
     }
@@ -16,15 +15,7 @@ export class BEVService {
     // Información del contrato
     async getContractBalance(){
         let contractBalance = (await this.contract.getContractBalance()).toNumber();        
-        return contractBalance;
-        /*
-        try{
-            let contractBalance = (await this.contract.getContractBalance()).toNumber();                    
-            return this.response(200, contractBalance);
-        } catch(e) {
-            console.error(e);
-            return this.response(500, e);
-        }*/        
+        return contractBalance;       
     }
 
     async getTotalElections() {
@@ -42,26 +33,46 @@ export class BEVService {
 
     // Información de una cuenta
     async isAdmin(addr) {
-        let isUserAdmin = await this.contract.admins(addr);
+        let isUserAdmin = await this.contract.admins(addr);        
         return isUserAdmin;
     }
-
-    // Solo el propietario del contrato puede agregar o quitar administradores
+    
     async addAdmin(address, account) {
-        let transactionInfo;
-        await this.contract.addAdmin(address, { from: account }).then((receipt) => {
-            transactionInfo = receipt;
-            });
-        return transactionInfo;
+        let isUserAdmin = await this.isAdmin(address);
+        if(isUserAdmin) {
+            return this.response(204,"El usuario ya es administrador");
+        }
+        else {
+            let transactionInfo;
+            await this.contract.addAdmin(address, { from: account }).then((receipt) => {                
+                transactionInfo = receipt;
+                });
+            return this.response(200,transactionInfo);
+        }
     }
 
     async deleteAdmin(address, account) {
-        console.log("bevService -> deleteAdmin");
-        let transactionInfo;
-        await this.contract.deleteAdmin(address, { from: account }).then((receipt) => {
-            transactionInfo = receipt;
-            });
-        return transactionInfo;
+        let isUserAdmin = await this.isAdmin(address);
+        if(!isUserAdmin) {
+            return this.response(204,"El usuario no es administrador");
+        }
+        else {
+            let transactionInfo;
+            await this.contract.deleteAdmin(address, { from: account }).then((receipt) => {
+                transactionInfo = receipt;
+                });
+            return this.response(200,transactionInfo);
+        }
+    }
+
+    structElection(election) {
+        return {
+            id: election[0].toNumber(),
+            name: election[1],
+            active: election[2].toString(),
+            candidatesCount: election[3].toNumber(),
+            votersCount: election[4].toNumber()
+        }
     }
 
     // Listado de elecciones
@@ -73,48 +84,60 @@ export class BEVService {
                 active: election[2].toString(),
                 candidatesCount: election[3].toNumber(),
                 votersCount: election[4].toNumber()
+                
             }
         })
     }
 
     async getElections() {
         let total = await this.getTotalElections();
-        let elections = [];        
-        for(var i = 1; i <= total; i++) {            
-            let election = await this.contract.getElection(i);
-            elections.push(election);
+        if(total > 0) {
+            let elections = [];        
+            for(var i = 1; i <= total; i++) {            
+                let election = await this.contract.getElection(i);
+                elections.push(election);
+            }
+            return this.response(200, this.mapElection(elections));
         }
-        return this.mapElection(elections);
-    }
+        else
+            return this.response(204, "No hay elecciones.");
 
-    // Ver una elección
-    async getElection(id) {        
-        let elections = [];            
-        let election = await this.contract.getElection(id);
-        elections.push(election);
-        return this.mapElection(elections);
     }
-
+    
     // Comprobar las elecciones en las que esta incluida la cuenta
     async getElectionsByAccount(account) {        
         let total = await this.getTotalElections();
-        let elections = [];        
-        for(var i = 1; i <= total; i++) {
-            if(await this.contract.voterIsJoined(i, account)) {
-                let _yaVoto = await this.voterHasVoted(i, account);
-                let e = await this.contract.getElection(i);                
-                let election = {
-                    id: e[0].toNumber(),
-                    name: e[1],
-                    active: e[2].toString(),
-                    candidatesCount: e[3].toNumber(),
-                    votersCount: e[4].toNumber(),
-                    yaVoto: _yaVoto.toString()
-                  };                                
-                elections.push(election);
+        if(total > 0) {
+            let elections = [];        
+            for(var i = 1; i <= total; i++) {
+                if(await this.contract.voterIsJoined(i, account)) {
+                    let _yaVoto = await this.voterHasVoted(i, account);
+                    let e = await this.contract.getElection(i);                
+                    let election = {
+                        id: e[0].toNumber(),
+                        name: e[1],
+                        active: e[2].toString(),
+                        candidatesCount: e[3].toNumber(),
+                        votersCount: e[4].toNumber(),
+                        yaVoto: _yaVoto.toString()
+                      };                                
+                    elections.push(election);
+                }
             }
+            return this.response(200,elections);
         }
-        return elections;
+        else 
+            return this.response(204, "No hay elecciones.");
+    }
+
+    // Ver una elección
+    async getElection(id) {                  
+        let election;
+        await this.contract.getElection(id).then((receipt) => {
+            election = this.structElection(receipt);
+        });
+        
+        return this.response(200, election);
     }
 
     // Agregar nueva elección y devuelve la información de la transacción
@@ -123,7 +146,7 @@ export class BEVService {
         await this.contract.addElection(name, { from: account, value: valueElection }).then((receipt) => {
             transactionInfo = receipt;
             });
-        return transactionInfo;
+        return this.response(200,transactionInfo);
     }
 
     // Eliminar una elección y devuelve la información de la transacción
@@ -132,7 +155,7 @@ export class BEVService {
         await this.contract.deleteElection(id, {from: account}).then((receipt) => {
             transactionInfo = receipt;
         });
-        return transactionInfo;
+        return this.response(200,transactionInfo);
     }
 
     // Listado de candidatos
@@ -149,33 +172,45 @@ export class BEVService {
 
     async getCandidates() {
         let totalElections = await this.getTotalElections();
-        let candidates = [];        
-        for(var i = 1; i <= totalElections; i++) {    
-            var totalCandidates = await this.contract.getCandidatesCount(i);
-            for(var j = 1 ; j <= totalCandidates; j++) {
-                let candidate = await this.contract.getCandidate(i,j);
-                candidates.push(candidate);
+        if(totalElections > 0) {
+            let candidates = [];        
+            for(var i = 1; i <= totalElections; i++) {    
+                let totalCandidates = await this.contract.getCandidatesCount(i);
+                for(var j = 1 ; j <= totalCandidates; j++) {
+                    let candidate = await this.contract.getCandidate(i,j);
+                    candidates.push(candidate);
+                }   
             }
+            if(candidates.length > 0)
+                return this.response(200, this.mapCandidate(candidates));
         }
-        return this.mapCandidate(candidates);
+            
+        return this.response(204, "No hay candidatos!");
     }
 
     async getCandidatesByElection(election) {        
         var totalCandidates = await this.contract.getCandidatesCount(election);
-        let candidates = [];                  
-        for(var i = 1 ; i <= totalCandidates; i++) {
-            let candidate = await this.contract.getCandidate(election,i);
-            candidates.push(candidate);
+        if(totalCandidates > 0) {
+            let candidates = [];                  
+            for(var i = 1 ; i <= totalCandidates; i++) {
+                await this.contract.getCandidate(election,i).then((candidate) => {
+                    candidates.push(candidate);
+                });
+            }
+            return this.response(200, this.mapCandidate(candidates));
         }
-        return this.mapCandidate(candidates);
+        else
+            return this.response(204, "No hay candidatos."); 
     }
 
     // Ver un candidato
-    async getCandidate(election, id) {        
-        let candidates = [];            
-        let candidate = await this.contract.getCandidate(election, id);
-        candidates.push(candidate);
-        return this.mapCandidate(candidates);
+    async getCandidate(election, id) {                  
+        let candidate;
+        await this.contract.getCandidate(election, id).then((receipt) => {
+            candidate = receipt
+        });
+        
+        return this.response(200, candidate);
     }
 
     // Agregar nuevo candidato y devuelve la información de la transacción
@@ -184,7 +219,7 @@ export class BEVService {
         await this.contract.addCandidate(election, name, { from: account }).then((receipt) => {
             transactionInfo = receipt;
             });
-        return transactionInfo;
+        return this.response(200, transactionInfo);
     }
 
     // Eliminar un candidato
@@ -193,7 +228,7 @@ export class BEVService {
         await this.contract.deleteCandidate(election, id, {from: account}).then((receipt) => {
             transactionInfo = receipt;
         });
-        return transactionInfo;
+        return this.response(200, transactionInfo);
     }
 
     // Listado de votantes
@@ -215,19 +250,21 @@ export class BEVService {
             var totalVoters = await this.contract.getVotersCount(i);
             for(var j = 1 ; j <= totalVoters; j++) {
                 let addr = await this.contract.getAddressVoter(i,j);
-                let voter = await this.contract.getVoter(i,addr);
+                let voter = await this.contract.getVoter(i,addr); // Busco el votante por su dirección.
                 voters.push(voter);
             }
         }
-        return this.mapVoter(voters);
+        return this.response(200, this.mapVoter(voters));
     }
 
     // Ver un votante
     async getVoter(election, address) {        
-        let voters = [];            
-        let voter = await this.contract.getVoter(election, address);
-        voters.push(voter);
-        return this.mapVoter(voters);
+        let voter;            
+        await this.contract.getVoter(election, address).then((receipt) => {
+            voter = receipt;
+        });
+        
+        return this.response(200, voter);
     }
 
     // Agregar nuevo votante y devuelve la información de la transacción
@@ -236,7 +273,7 @@ export class BEVService {
         await this.contract.addVoter(election, address, name, { from: account }).then((receipt) => {
             transactionInfo = receipt;
             });
-        return transactionInfo;
+        return this.response(200,transactionInfo);
     }
 
     // Eliminar un votante
@@ -245,7 +282,7 @@ export class BEVService {
         await this.contract.deleteVoter(election, address, {from: account}).then((receipt) => {
             transactionInfo = receipt;
         });
-        return transactionInfo;
+        return this.response(200, transactionInfo);
     }
 
     // Consulto si el usuario ya voto
@@ -260,27 +297,16 @@ export class BEVService {
         await this.contract.voting(election, candidate, {from: account}).then((receipt) => {
             transactionInfo = receipt;
         });
-        
-        // TODO: addTransaction -> Ver si se puede integrar en el contrato.
-        // No hace falta ya que en MetaMask se guardar los id de transacciones. Alcanzaria solo con mostrarlo
-        await this.contract.addTransaction(election, transactionInfo.tx, {from: account}).then((receipt) => {
-            console.log("Transaccion registrada!");
-        });
 
-        return transactionInfo;
+        return this.response(200, transactionInfo);
     }
     
-    // Obtengo la información de la transacción
-    async getTransaction(election, account){
-        let transaction = await this.contract.getTransaction(election, account);
-        return transaction;
-    }
-
     // Resultado de votación
     async getResultElection(election) {
-        let ganador = await this.contract.getResultElection(election);
-        let candidatoGanador = await this.getCandidate(election, ganador);
-        return candidatoGanador;
+        let id = await this.contract.getResultElection(election);
+        let candidatoGanador = await this.contract.getCandidate(election, id);
+        
+        return this.response(200,candidatoGanador);
     }
     
     // Transferencia de fondos del contrato
@@ -289,7 +315,7 @@ export class BEVService {
         await this.contract.transferFromContract(address, etherToRefund, {from: account}).then((receipt) => {
             transactionInfo = receipt;
         });
-        return transactionInfo;
+        return this.response(200, transactionInfo);
     }
 
 }
